refactor: split KeyboardHandler into input listener and text output (#35)
- Extract HotkeyListener to src/dicton/input/hotkey_listener.py
- Extract TextOutputService to src/dicton/output/text_output.py
- Delete keyboard_handler.py
- Wire SessionService to TextOutputService, RuntimeService to HotkeyListener
- Update tests to use the new classes
