+++ src/dicton/input/__init__.py
@@ -1,0 +1,1 @@
+"""Keyboard input handling for Dicton."""
