+++ src/dicton/output/__init__.py
@@ -1,0 +1,1 @@
+"""Text output handling for Dicton."""
