+++ tests/test_hotkey.py
@@ -10,13 +10,13 @@ Key = pynput.keyboard.Key
 
 
 def reload_modules():
-    """Reload config and keyboard_handler to pick up env changes."""
+    """Reload config and hotkey_listener to pick up env changes."""
     import dicton.config as config_module
-    import dicton.keyboard_handler as kb_module
+    import dicton.input.hotkey_listener as hl_module
 
     importlib.reload(config_module)
-    importlib.reload(kb_module)
-    return kb_module
+    importlib.reload(hl_module)
+    return hl_module
 
 
 class TestHotkeyParsing:
@@ -27,8 +27,8 @@ class TestHotkeyParsing:
         monkeypatch.setenv("HOTKEY_MODIFIER", "alt")
         monkeypatch.setenv("HOTKEY_KEY", "g")
 
-        kb_module = reload_modules()
-        handler = kb_module.KeyboardHandler(on_toggle_callback=lambda: None)
+        hl_module = reload_modules()
+        handler = hl_module.HotkeyListener(on_toggle_callback=lambda: None)
         handler.pressed_keys = {Key.alt, "g"}
 
         assert handler._is_hotkey_pressed() is True
@@ -38,8 +38,8 @@ class TestHotkeyParsing:
         monkeypatch.setenv("HOTKEY_MODIFIER", "ctrl")
         monkeypatch.setenv("HOTKEY_KEY", "h")
 
-        kb_module = reload_modules()
-        handler = kb_module.KeyboardHandler(on_toggle_callback=lambda: None)
+        hl_module = reload_modules()
+        handler = hl_module.HotkeyListener(on_toggle_callback=lambda: None)
         handler.pressed_keys = {Key.ctrl, "h"}
 
         assert handler._is_hotkey_pressed() is True
@@ -49,8 +49,8 @@ class TestHotkeyParsing:
         monkeypatch.setenv("HOTKEY_MODIFIER", "alt")
         monkeypatch.setenv("HOTKEY_KEY", "g")
 
-        kb_module = reload_modules()
-        handler = kb_module.KeyboardHandler(on_toggle_callback=lambda: None)
+        hl_module = reload_modules()
+        handler = hl_module.HotkeyListener(on_toggle_callback=lambda: None)
         handler.pressed_keys = {Key.alt_l, "g"}
 
         assert handler._is_hotkey_pressed() is True
@@ -60,8 +60,8 @@ class TestHotkeyParsing:
         monkeypatch.setenv("HOTKEY_MODIFIER", "alt")
         monkeypatch.setenv("HOTKEY_KEY", "g")
 
-        kb_module = reload_modules()
-        handler = kb_module.KeyboardHandler(on_toggle_callback=lambda: None)
+        hl_module = reload_modules()
+        handler = hl_module.HotkeyListener(on_toggle_callback=lambda: None)
         handler.pressed_keys = {"g"}  # Only key, no modifier
 
         assert handler._is_hotkey_pressed() is False
@@ -71,8 +71,8 @@ class TestHotkeyParsing:
         monkeypatch.setenv("HOTKEY_MODIFIER", "alt")
         monkeypatch.setenv("HOTKEY_KEY", "g")
 
-        kb_module = reload_modules()
-        handler = kb_module.KeyboardHandler(on_toggle_callback=lambda: None)
+        hl_module = reload_modules()
+        handler = hl_module.HotkeyListener(on_toggle_callback=lambda: None)
         handler.pressed_keys = {Key.alt}  # Only modifier, no key
 
         assert handler._is_hotkey_pressed() is False
@@ -82,33 +82,33 @@ class TestHotkeyParsing:
         monkeypatch.setenv("HOTKEY_MODIFIER", "alt")
         monkeypatch.setenv("HOTKEY_KEY", "G")  # Uppercase
 
-        kb_module = reload_modules()
-        handler = kb_module.KeyboardHandler(on_toggle_callback=lambda: None)
+        hl_module = reload_modules()
+        handler = hl_module.HotkeyListener(on_toggle_callback=lambda: None)
         handler.pressed_keys = {Key.alt, "g"}  # lowercase
 
         assert handler._is_hotkey_pressed() is True
 
 
-class TestKeyboardHandlerInit:
-    """Test KeyboardHandler initialization."""
+class TestHotkeyListenerInit:
+    """Test HotkeyListener initialization."""
 
     def test_init_with_callback(self):
-        """Test handler initializes with callback."""
-        kb_module = reload_modules()
+        """Test listener initializes with callback."""
+        hl_module = reload_modules()
 
         callback_called = []
 
         def callback():
             callback_called.append(True)
 
-        handler = kb_module.KeyboardHandler(on_toggle_callback=callback)
+        handler = hl_module.HotkeyListener(on_toggle_callback=callback)
         assert handler.on_toggle is callback
         assert handler.listener is None
         assert len(handler.pressed_keys) == 0
 
     def test_init_without_callback(self):
-        """Test handler initializes without callback."""
-        kb_module = reload_modules()
+        """Test listener initializes without callback."""
+        hl_module = reload_modules()
 
-        handler = kb_module.KeyboardHandler(on_toggle_callback=None)
+        handler = hl_module.HotkeyListener(on_toggle_callback=None)
         assert handler.on_toggle is None
